# static/innovasoft_openapi_cache.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ApiPruebaReactJs",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/Api"
    }
  ],
  "paths": {
    "/api/Authenticate/login": {
      "post": {
        "tags": [
          "Authenticate"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Authenticate/register": {
      "post": {
        "tags": [
          "Authenticate"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cliente/Listado": {
      "post": {
        "tags": [
          "Cliente"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListarCliente"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ListarCliente"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ListarCliente"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cliente_DTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cliente_DTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cliente_DTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Cliente/Crear": {
      "post": {
        "tags": [
          "Cliente"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteCrear"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteCrear"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteCrear"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cliente/Actualizar": {
      "post": {
        "tags": [
          "Cliente"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteActualizar"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteActualizar"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteActualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cliente/Obtener/{IdCliente}": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "parameters": [
          {
            "name": "IdCliente",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleCliente_DTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleCliente_DTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleCliente_DTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Cliente/Eliminar/{IdCliente}": {
      "delete": {
        "tags": [
          "Cliente"
        ],
        "parameters": [
          {
            "name": "IdCliente",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Intereses/Listado": {
      "get": {
        "tags": [
          "Intereses"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Interes_DTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Interes_DTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Interes_DTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ClienteActualizar": {
        "required": [
          "apellidos",
          "celular",
          "direccion",
          "fAfiliacion",
          "fNacimiento",
          "id",
          "identificacion",
          "interesFK",
          "nombre",
          "otroTelefono",
          "sexo",
          "usuarioId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string"
          },
          "apellidos": {
            "type": "string"
          },
          "identificacion": {
            "type": "string"
          },
          "celular": {
            "type": "string"
          },
          "otroTelefono": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "fNacimiento": {
            "type": "string",
            "format": "date-time"
          },
          "fAfiliacion": {
            "type": "string",
            "format": "date-time"
          },
          "sexo": {
            "type": "string"
          },
          "resennaPersonal": {
            "type": "string",
            "nullable": true
          },
          "imagen": {
            "type": "string",
            "nullable": true
          },
          "interesFK": {
            "type": "string",
            "format": "uuid"
          },
          "usuarioId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ClienteCrear": {
        "required": [
          "apellidos",
          "celular",
          "direccion",
          "fAfiliacion",
          "fNacimiento",
          "identificacion",
          "interesFK",
          "nombre",
          "otroTelefono",
          "sexo",
          "usuarioId"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "apellidos": {
            "type": "string"
          },
          "identificacion": {
            "type": "string"
          },
          "celular": {
            "type": "string"
          },
          "otroTelefono": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "fNacimiento": {
            "type": "string",
            "format": "date-time"
          },
          "fAfiliacion": {
            "type": "string",
            "format": "date-time"
          },
          "sexo": {
            "type": "string"
          },
          "resennaPersonal": {
            "type": "string",
            "nullable": true
          },
          "imagen": {
            "type": "string",
            "nullable": true
          },
          "interesFK": {
            "type": "string",
            "format": "uuid"
          },
          "usuarioId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "Cliente_DTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "identificacion": {
            "type": "string",
            "nullable": true
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellidos": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DetalleCliente_DTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellidos": {
            "type": "string",
            "nullable": true
          },
          "identificacion": {
            "type": "string",
            "nullable": true
          },
          "telefonoCelular": {
            "type": "string",
            "nullable": true
          },
          "otroTelefono": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "fNacimiento": {
            "type": "string",
            "format": "date-time"
          },
          "fAfiliacion": {
            "type": "string",
            "format": "date-time"
          },
          "sexo": {
            "type": "string"
          },
          "resenaPersonal": {
            "type": "string",
            "nullable": true
          },
          "imagen": {
            "type": "string",
            "nullable": true
          },
          "interesesId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "Interes_DTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ListarCliente": {
        "required": [
          "usuarioId"
        ],
        "type": "object",
        "properties": {
          "identificacion": {
            "type": "string",
            "nullable": true
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "usuarioId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "LoginModel": {
        "required": [
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "RegisterModel": {
        "required": [
          "email",
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "WeatherForecast": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "temperatureC": {
            "type": "integer",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}